lotta updates
